moved TasksTypes to Task utils and renamed it to taskFilterTabs
